Implement avatar, user can upload avatar when signup
User now can upload avatar when signup
Player and PlayerController has modified that all numerical fields are
in Integer type, not Number.
When user uploading an avatar by signup, it would be too large to pass
image locally without opening up gallary. Instead, avatar is uploaded to
firestore and leaderboard or some other Activity need to fetch it when
necessary.